requirements.txt
torch==2.2.0
transformers==4.36.2
pdfminer.six==20221105
numpy==1.24.3
tqdm==4.65.0
scikit-learn==1.3.2 